+++ br/com/system/view/parameter/list.php
@@ -83,9 +83,11 @@
                         echo '<a title="Ativar dados!" href="', server_url('?page=ControllerParameter&option=enable&param_pk_id=' . $each_parameters->para_pk_id), '" class="btn btn-success btn-circle btn-sm excluir" style="margin: 5px">';
                         echo '<i class="fas fa-check-circle"></i>';
                         echo '</a>';
-                        echo '<a title="Excluir dados!" href="', server_url('?page=ControllerParameter&option=delete&param_pk_id=' . $each_parameters->para_pk_id), '" class="btn btn-danger btn-circle btn-sm excluir" onclick="return confirm(´Deseja realmente excluir, esta operação não podera ser desfeita!´)" style="margin: 5px">';
-                        echo '<i class="fas fa-trash"></i>';
-                        echo '</a>';
+                        if ($permissao == 1) {
+                            echo '<a title="Excluir dados!" href="', server_url('?page=ControllerParameter&option=delete&param_pk_id=' . $each_parameters->para_pk_id), '" class="btn btn-danger btn-circle btn-sm excluir" onclick="return confirm(´Deseja realmente excluir, esta operação não podera ser desfeita!´)" style="margin: 5px">';
+                            echo '<i class="fas fa-trash"></i>';
+                            echo '</a>';
+                        }
                     }
                     echo '</td>';
                     echo '</tr>';
